+++ config/initializers/assets.rb
@@ -7,6 +7,9 @@ Rails.application.config.assets.precompile += %w[
   component_guide/filter-components.js
   component_guide/visual-regression.js
   component_guide/print.css
+  govuk_publishing_components/rum-loader.js
+  govuk_publishing_components/vendor/lux.js
+  govuk_publishing_components/vendor/lux-polyfill.js
   govuk_publishing_components/all_components.js
   govuk_publishing_components/ie.js
   govuk_publishing_components/modules.js